pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.1",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "discord-py>=2.6.4",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "matplotlib>=3.10.7",
    "newsapi>=0.1.1",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "praw>=7.8.1",
    "pytelegrambotapi>=4.29.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "textblob>=0.19.0",
    "tweepy>=4.16.0",
    "yfinance>=0.2.66",
]
